Add differential fuzzer
This tool runs by comparing minifier-rs output with rust-cssparser.
It follows a "don't make promises you can't keep" rule where,
if either parser fails, we ignore the case. This might not be what
you expect, but it means we don't have to support every corner of
the CSS syntax, as long as we're smart enough to fail when we hit
something we can't handle.